script updates for windows os

+++ install-scripts/blacksmith-cli-install-win-x64.ps1
@@ -10,11 +10,8 @@ Write-Host "Installing $toInstallVersion..."
 
 # Define variables
 $platform = "win-x64"
-# $localAppFolder = "C:\Program Files\Blacksmith"
 $blacksmithAppFolder = "C:\Program Files\Blacksmith"
-#$blacksmithCliApp = "$blacksmithAppFolder\blacksmith-cli.exe"
-
-
+$blacksmithAppFolderOld = "C:\Program Files\Blacksmith-Old"
 $url = "https://github.com/blacksmithdata/blacksmith-cli/releases/download/$toInstallVersion/$toInstallVersion-$platform.zip"
 $blacksmithInstallFolder = Join-Path $env:USERPROFILE "Downloads\blacksmith-cli-install-temp"
 
@@ -23,7 +20,10 @@ Write-Host "----- BEGIN INSTALLATION ...."
 # STEP 1: CLEANUP INSTALL FOLDER
 Write-Host ""
 Write-Host "----- STEP 1: CLEANUP INSTALL FOLDER"
-Remove-Item -Path $blacksmithInstallFolder -Recurse -Force
+if (Test-Path $blacksmithInstallFolder) {
+  Remove-Item -Path $blacksmithInstallFolder -Recurse -Force
+}
+
 New-Item -Path $blacksmithInstallFolder -ItemType Directory | Out-Null
 Set-Location -Path $blacksmithInstallFolder
 
@@ -43,28 +43,42 @@ Rename-Item -Path $toInstallVersion -NewName "blacksmith"
 # STEP 4: PREPARE CLI DESTINATION FOLDER
 Write-Host ""
 Write-Host "----- STEP 4: PREPARE CLI DESTINATION FOLDER"
-Move-Item -Path $blacksmithAppFolder -Destination "C:\Program Files\blacksmith-old" -Force
+if (Test-Path $blacksmithAppFolder) {
+  Move-Item -Path $blacksmithAppFolder -Destination $blacksmithAppFolderOld -Force
+}
+
 Move-Item -Path "blacksmith" -Destination $blacksmithAppFolder -Force
-New-Item -Path $blacksmithAppFolder -ItemType Directory | Out-Null
+# New-Item -Path $blacksmithAppFolder -ItemType Directory | Out-Null
 ###########################
 
 ###########################
 # STEP 5: DELETE CACHE FOLDER
 Write-Host ""
 Write-Host "----- STEP 5: DELETE CACHE FOLDER"
-Remove-Item -Path "C:\Program Files\blacksmith-old" -Recurse -Force
-Set-Location -Path $env:USERPROFILE
+if (Test-Path $blacksmithAppFolderOld) {
+  Remove-Item -Path $blacksmithAppFolderOld -Recurse -Force
+}
+
+Set-Location -Path $env:USERPROFILE\"Downloads"
 Remove-Item -Path $blacksmithInstallFolder -Recurse -Force
 ###########################
 
 ###########################
-# STEP 6: APPLY SYMLINK (NOT APPLICABLE FOR WINDOWS OS)
-# Write-Host ""
-# Write-Host "----- STEP 6: APPLY SYMLINK"
-# Remove-Item -Path $blacksmithAppBin -Force
-# New-Item -ItemType SymbolicLink -Path $blacksmithAppBin -Target "$blacksmithAppFolder\blacksmith-cli.exe"
+# STEP 6: APPLY PATH IN ENV VARIABLES FOR THE LOGGEDIN USER
+$currentPath = [System.Environment]::GetEnvironmentVariable("Path", [System.EnvironmentVariableTarget]::User)
+
+# Check if the folder exists in the Path variable
+if ($currentPath -like "*$blacksmithAppFolder*") {    
+    Write-Host "Path $blacksmithAppFolder already exists."
+} else {
+    Write-Host "Path $blacksmithAppFolder does not exists, adding it..."
+    $newPath = "$currentPath;$blacksmithAppFolder"
+    [System.Environment]::SetEnvironmentVariable("Path", $newPath, [System.EnvironmentVariableTarget]::User)
+    Write-Host "Path $blacksmithAppFolder added to user's path..."
+}
 ###########################
 
 Write-Host ""
-Write-Host "Install completed. To get started, type blacksmith-cli"
+Write-Host "Install completed. Close all your terminal and launch again."
+Write-Host "To get started, type blacksmith-cli or blacksmith-cli.exe"
 Write-Host ""
